Add ApIJenkinsViewTest with new tests
